Retire VNNI dialect
Retire VNNI dialect and VNNI TPP operations.

History: We introduced the VNNI dialect before TPP on tensor. Tensor
semantics was needed as packing happens at this abstraction. Now that we
have TPP on tensor, we no longer need the VNNI dialect.

Annoyance: We use an explicit VNNI layout (i.e., the layout is reflected
on the tensor/memef shape); this complicates the verification for the
ops and the lowering to XSMM and Loops. Ideally, the layout should be
metadata we attach on the tensor/memref. For example, we could write
something like:

`memref<32x32xbf16, affine_map<(d0, d1) -> (d0/VNNI, d1,VNNI)>`

This is possible today, but to relayout a tensor, we use
`tensor.pack`, which makes the layout explicit in the tensor shape.