Make `linalg.transpose` 2d.
The idea is to tile the non-interchanged dimensions. But this can
prevent mapping to tpp as the stride on the fasted-varying dimension may
not be one anymore.

This is interesting. We map at tensor but we don't have any notion of
strides, then we bufferize. During bufferization we may have innermost
strides not equal to 1, thus we cannot map to XSMM anymore.

Note use some improvement in LLVM - canonicalize transpose operation if
identity.